+++ SPC-2021/api/client/client/Client.Logic/Models/Client Models/EditClientAssignmentStatusModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Santa.Api.Models.Client_Models
+namespace SharkTank.Logic.Client_Models
 {
     public class EditClientAssignmentStatusModel
     {
