+++ SPC-2021/api/client/client/Client.Logic/Models/Client Models/EditClientNameModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Santa.Api.Models
+namespace Santa.Logic.Client_Models
 {
     public class EditClientNameModel
     {
