+++ SPC-2021/api/client/client/Client.Logic/Models/Client Models/EditNewClientModel.cs
@@ -1,12 +1,9 @@
-﻿using Santa.Api.Models.Survey_Response_Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Santa.Api.Models.Client_Models
+namespace Santa.Logic.Client_Models
 {
-    public class NewClientWithResponsesModel
+    public class EditNewClientModel
     {
         public Guid clientStatusID { get; set; }
         public string clientName { get; set; }
@@ -18,8 +15,5 @@ namespace Santa.Api.Models.Client_Models
         public string clientState { get; set; }
         public string clientPostalCode { get; set; }
         public string clientCountry { get; set; }
-        public List<ApiSurveyResponse> responses { get; set; }
-        public bool isAdmin { get; set; }
-        public bool hasAccount { get; set; }
     }
 }
